requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
mysql-connector-python==9.2.0
Flask==3.1.0
